Improved class hierarchy.
Added base class for allocator parameters and derived the BnB and ILP parameter classes from here.

VMAllocator is now an interface. The BnB and ILP allocators implement this interface.

Updated the config parser to handle the polymorphic parameter classes.

ILP solvers are configured in the config file similarly to BnB solvers.

Updated the sample config file.