Cleaned up VM and PM ID usage.
The IDs are not used in the BnB algorithm directly anymore.

The allocation is now a hash map with VMs as keys and PMs as values.

The non-fitting VMs are stored in Change.h directly.

The initialPM variable of VMs is set in the constructor and used afterwards for comparisons.

The IDs, however, are used for printing allocations (according to ascending order of VM IDs).

Other changes:
- Improved the VMAllocator interface
- Best cost is now uniformly -1 when no valid allocation was found
- Moved parameter configuration logging to main.cpp and improved log formatting
- Removed global_params.h (is not required since revision 17)
- Renamed some variables and functions for better understanding